+++ CHANGES.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+1.3.1 - 2023-10-28
+------------------
+
+- Modified existing ``setup.py`` hook from an install hook to a build
+  hook to ensure the generated module files are present.  Should any of
+  those modules are missing and the required dependencies for are not
+  present (i.e. ``ply`` and ``setuptools``), the build will result in a
+  non-zero exit status and the documented error message should reflect
+  which of the required dependencies are missing.
+
 1.3.0 - 2021-10-08
 ------------------